requirements.txt
# os
# json
requests
pandas 
plotly.express
streamlit